requirements.txt
apache-airflow
astro-sdk-python
pyspark
minio
pandas
python-dotenv